+++ src/ashley/machina_extensions/forum_conversation/__init__.py
@@ -1,0 +1,6 @@
+"""This application overrides to django-machina's forum_conversation app."""
+
+# pylint: disable=invalid-name
+default_app_config = (
+    "ashley.machina_extensions.forum_conversation.apps.ForumConversationAppConfig"
+)