+++ src/ashley/machina_extensions/forum_conversation/models.py
@@ -1,0 +1,4 @@
+"""Declare the models related to the forum_conversation app ."""
+
+# Import models from the django-machina app
+from machina.apps.forum_conversation.models import *  # noqa, pylint: disable=unused-wildcard-import,wildcard-import